make weapon switching allow weapon names
- prepare netcode files for server-sided behaviours
- knockback on players stop grounded speed capping (changed a weapon param)
- adjust player camera behaviour to prevent clipping through geometry
- renamed WeaponBulletMgr to WeaponBase for consistency
