+++ Gunball/Assets/Resources/Prefabs/BulletTypes/Blast.prefab
@@ -26,7 +26,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 4266941966911170520}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -26.377724, y: -9.360657, z: -70.498474}
+  m_LocalPosition: {x: -33.493652, y: -1.4371083, z: -51.582916}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 5283629536417723090}
