+++ Gunball/Assets/Resources/Prefabs/Weapons/WeaponGuideDebug.prefab
@@ -27,7 +27,6 @@ Transform:
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
@@ -88,7 +87,7 @@ MonoBehaviour:
     Knockback:
       Force: 4
       Radius: 0
-      DistanceBias: 0
+      TimeBias: 0.05
       PlayerBias: 0.2
       VsBallBias: 1
       MapObjectBias: 0.25
