+++ Gunball/Assets/Scripts/MapObject/IShootableObject.cs
@@ -17,6 +17,7 @@ namespace Gunball.MapObject
         Transform Transform { get; }
         ShootableType Type { get; }
         public void Knockback(Vector3 force, Vector3 pos);
+        public void SetKnockbackTimer(float timeBias);
         public void DoDamage(float damage, Player source = null);
         public void RecoverDamage(float healing, Player source = null);
