+++ Gunball/Assets/Scripts/NetPlay.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 80ef566280cae15418e8370b8b68a256
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
