+++ Gunball/Assets/Scripts/NetPlay/NetworkedGunball.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2ce768978d134046ad3651fcc2a39af
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
