+++ Gunball/Assets/Scripts/NetPlay/NetworkedScoringSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 928f3959351224d4c8303374740b1c9a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
