--- Gunball/Assets/Scripts/Player/NetworkedPlayer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a22c8cd7a4ea047418625e6cff2e1a5c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
