+++ Gunball/Assets/Scripts/WeaponBullet/WeaponBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 549134579bb10cb46bed3e8ba0872e3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
